exporter/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
dotenv==0.9.9
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.1.0
javaobj-py3==0.4.4
packaging==25.0
pluggy==1.5.0
prometheus_client==0.21.1
psycopg2-binary==2.9.10
pytest==8.3.5
python-dotenv==1.1.0
python-gitlab==6.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
tomli==2.2.1
urllib3==2.4.0
uv==0.7.2
